dorothy_personel/requirements.txt
# requirements.txt
dlib==19.24.2
joblib==1.3.2
numpy==1.26.1
scikit-learn==1.3.0
pillow